+++ driver/src/main/java/org/neo4j/driver/Config.java
@@ -41,6 +41,7 @@ import org.neo4j.driver.reactive.RxSession;
 import org.neo4j.driver.util.Immutable;
 import org.neo4j.driver.util.Resource;
 
+import static java.lang.String.format;
 import static org.neo4j.driver.Logging.javaUtilLogging;
 
 /**
@@ -283,7 +284,7 @@ public class Config
         private long idleTimeBeforeConnectionTest = PoolSettings.DEFAULT_IDLE_TIME_BEFORE_CONNECTION_TEST;
         private long maxConnectionLifetimeMillis = PoolSettings.DEFAULT_MAX_CONNECTION_LIFETIME;
         private long connectionAcquisitionTimeoutMillis = PoolSettings.DEFAULT_CONNECTION_ACQUISITION_TIMEOUT;
-        private String userAgent = ConnectionSettings.DEFAULT_USER_AGENT;
+        private String userAgent = format( "neo4j-java/%s", driverVersion() );
         private final SecuritySettings.SecuritySettingsBuilder securitySettingsBuilder = new SecuritySettings.SecuritySettingsBuilder();
         private int routingFailureLimit = RoutingSettings.DEFAULT.maxRoutingFailures();
         private long routingRetryDelayMillis = RoutingSettings.DEFAULT.retryTimeoutDelay();
@@ -752,6 +753,25 @@ public class Config
             }
             this.userAgent = userAgent;
             return this;
+        }
+
+        /**
+         * Extracts the driver version from the driver jar MANIFEST.MF file.
+         */
+        private static String driverVersion()
+        {
+            // "Session" is arbitrary - the only thing that matters is that the class we use here is in the
+            // 'org.neo4j.driver' package, because that is where the jar manifest specifies the version.
+            // This is done as part of the build, adding a MANIFEST.MF file to the generated jarfile.
+            Package pkg = Session.class.getPackage();
+            if ( pkg != null && pkg.getImplementationVersion() != null )
+            {
+                return pkg.getImplementationVersion();
+            }
+
+            // If there is no version, we're not running from a jar file, but from raw compiled class files.
+            // This should only happen during development, so call the version 'dev'.
+            return "dev";
         }
 
         /**
